+++ docs/CHANGELOG.md
@@ -29,6 +29,7 @@ General:
    - Added a new lua script to slice a node into smaller pieces
    - Expose shape generators to lua
    - Fixed color intensity handling for `kvx` files
+   - Added write support for `kvx` format (used in e.g. voxel doom and eduke3d)
 
 VoxEdit: